L23. Added BottomNavigationView on ActivityC.
