+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <activity
             android:name=".ActivityC"
             android:exported="false"
-            android:theme="@style/Theme.AppCompat.DayNight"
+            android:theme="@style/Theme.MaterialComponents.DayNight"
             android:launchMode="singleTask" />
         <activity
             android:name=".ActivityA"
