+++ app/src/main/res/layout/fragment_profile.xml
@@ -1,0 +1,173 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ProfileFragment">
+
+    <Spinner
+        android:id="@+id/spinner"
+        android:layout_width="@dimen/zero_dp"
+        android:layout_height="wrap_content"
+        android:entries="@array/genders"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/textView5"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/textView5"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:paddingStart="@dimen/general_margin"
+        android:text="@string/gender"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintBaseline_toBaselineOf="@+id/spinner"
+        app:layout_constraintEnd_toStartOf="@+id/spinner"
+        app:layout_constraintStart_toEndOf="@+id/et_birthdate" />
+
+    <EditText
+        android:id="@+id/et_birthdate"
+        android:layout_width="@dimen/zero_dp"
+        android:layout_height="wrap_content"
+        android:hint="@string/birthdate"
+        android:inputType="text"
+        app:layout_constraintBaseline_toBaselineOf="@+id/textView5"
+        app:layout_constraintEnd_toStartOf="@+id/textView5"
+        app:layout_constraintHorizontal_chainStyle="spread"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <TextView
+        android:id="@+id/tv_experience"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/middle_margin"
+        android:text="@string/tv_experience_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/btn_get_location" />
+
+    <RadioGroup
+        android:id="@+id/rg_experience"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/tv_experience">
+
+        <RadioButton
+            android:id="@+id/rb_s_experience"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:buttonTint="@color/pink_dark"
+            android:checked="true"
+            android:text="@string/rb_s_experience_text"
+            android:textSize="@dimen/header2_size" />
+
+        <RadioButton
+            android:id="@+id/rb_m_experience"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/little_margin"
+            android:buttonTint="@android:color/holo_blue_dark"
+            android:text="@string/rb_m_experience_text"
+            android:textSize="@dimen/header2_size" />
+
+        <RadioButton
+            android:id="@+id/rb_l_experience"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:buttonTint="@android:color/holo_green_dark"
+            android:text="@string/rb_l_experience_text"
+            android:textSize="@dimen/header2_size" />
+
+    </RadioGroup>
+
+    <TextView
+        android:id="@+id/tv_hobby"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/middle_margin"
+        android:text="@string/tv_hobby_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/rg_experience" />
+
+    <CheckBox
+        android:id="@+id/cb_football"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:buttonTint="@color/pink_light"
+        android:text="@string/cb_football_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tv_hobby" />
+
+    <CheckBox
+        android:id="@+id/cb_volleyball"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginHorizontal="@dimen/little_margin"
+        android:buttonTint="@android:color/holo_blue_light"
+        android:text="@string/cb_volleyball_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toEndOf="@+id/cb_football"
+        app:layout_constraintTop_toBottomOf="@id/tv_hobby" />
+
+    <CheckBox
+        android:id="@+id/cb_basketball"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:buttonTint="@android:color/holo_green_light"
+        android:text="@string/cb_basketball_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toEndOf="@+id/cb_volleyball"
+        app:layout_constraintTop_toBottomOf="@id/tv_hobby" />
+
+    <TextView
+        android:id="@+id/tv_bench_press"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/middle_margin"
+        android:text="@string/tv_bench_press"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/cb_football" />
+
+    <Spinner
+        android:id="@+id/sp_bench_press"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/middle_margin"
+        app:layout_constraintBottom_toBottomOf="@id/tv_bench_press"
+        app:layout_constraintStart_toEndOf="@id/tv_bench_press"
+        app:layout_constraintTop_toTopOf="@id/tv_bench_press" />
+
+    <TextView
+        android:id="@+id/tv_city"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/tv_city_text"
+        android:textSize="@dimen/header2_size"
+        app:layout_constraintBottom_toBottomOf="@id/et_city"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/et_birthdate" />
+
+    <EditText
+        android:id="@+id/et_city"
+        android:layout_width="@dimen/zero_dp"
+        android:layout_height="wrap_content"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/tv_city"
+        app:layout_constraintTop_toBottomOf="@+id/et_birthdate" />
+
+    <Button
+        android:id="@+id/btn_get_location"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/btn_get_location_text"
+        android:textSize="@dimen/header3_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/et_city" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
