+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/mi_profile"
+        android:icon="@drawable/ic_profile"
+        android:title="Профиль" />
+
+    <item
+        android:id="@+id/mi_work"
+        android:icon="@drawable/ic_work"
+        android:title="Работа" />
+
+    <item
+        android:id="@+id/mi_hobby"
+        android:icon="@drawable/ic_hobby"
+        android:title="Хобби" />
+
+</menu>
