+++ app/src/main/res/values/dimens.xml
@@ -11,8 +11,8 @@
     <dimen name="image_padding_end">10dp</dimen>
     <dimen name="image_padding_bottom">20dp</dimen>
     <dimen name="zero_dp">0dp</dimen>
-    <dimen name="left_guideline">109dp</dimen>
-    <dimen name="right_guideline">263dp</dimen>
+    <dimen name="left_guideline">115dp</dimen>
+    <dimen name="right_guideline">115dp</dimen>
     <dimen name="btn_toast_margin_top">70dp</dimen>
     <dimen name="tv_general_margin">5dp</dimen>
     <dimen name="et_layout_margin_top">10dp</dimen>
@@ -31,4 +31,5 @@
     <dimen name="border_radius">5dp</dimen>
     <dimen name="divider_height">0.5dp</dimen>
     <dimen name="rv_friends_padding">1dp</dimen>
+    <dimen name="bottom_navigation_view_height">75dp</dimen>
 </resources>
