+++ app/src/main/res/values/strings.xml
@@ -41,6 +41,7 @@
     <string name="btn_get_location_text">Получить местоположение</string>
     <string name="tv_friends_text">Друзья:</string>
     <string name="friend_item_tv_name_text">Name Template</string>
+    <string name="tv_fragment_hobby_text">Хоббихорсинг - тоже спорт!</string>
 
     <string-array name="genders">
         <item>Мужской</item>
